backend/requirements.txt
djangorestframework==3.6.3
uuid==1.30
